+++ lib/honeybadger/events_worker.rb
@@ -137,9 +137,13 @@ module Honeybadger
     def kill!
       d { "killing worker thread" }
 
+      if timeout_thread
+        Thread.kill(timeout_thread)
+        timeout_thread.join # Allow ensure blocks to execute.
+      end
+
       if thread
         Thread.kill(thread)
-        Thread.kill(timeout_thread)
         thread.join # Allow ensure blocks to execute.
       end