+++ lib/honeybadger/gauge.rb
@@ -19,11 +19,11 @@ module Honeybadger
     def payloads
       [
         {
-          total: @total,
-          min: @min,
-          max: @max,
-          avg: @avg,
-          latest: @latest
+          total: @total&.round(2),
+          min: @min&.round(2),
+          max: @max&.round(2),
+          avg: @avg&.round(2),
+          latest: @latest&.round(2)
         }
       ]
     end